refactor(authentication): Authentication module removed
