+++ angular.json
@@ -292,46 +292,6 @@
           }
         }
       }
-    },
-    "@knora/authentication": {
-      "root": "projects/knora/authentication",
-      "sourceRoot": "projects/knora/authentication/src",
-      "projectType": "library",
-      "prefix": "kui",
-      "architect": {
-        "build": {
-          "builder": "@angular-devkit/build-ng-packagr:build",
-          "options": {
-            "tsConfig": "projects/knora/authentication/tsconfig.lib.json",
-            "project": "projects/knora/authentication/ng-package.json"
-          },
-          "configurations": {
-            "production": {
-              "project": "projects/knora/authentication/ng-package.prod.json"
-            }
-          }
-        },
-        "test": {
-          "builder": "@angular-devkit/build-angular:karma",
-          "options": {
-            "main": "projects/knora/authentication/src/test.ts",
-            "tsConfig": "projects/knora/authentication/tsconfig.spec.json",
-            "karmaConfig": "projects/knora/authentication/karma.conf.js"
-          }
-        },
-        "lint": {
-          "builder": "@angular-devkit/build-angular:tslint",
-          "options": {
-            "tsConfig": [
-              "projects/knora/authentication/tsconfig.lib.json",
-              "projects/knora/authentication/tsconfig.spec.json"
-            ],
-            "exclude": [
-              "**/node_modules/**"
-            ]
-          }
-        }
-      }
     }
   },
   "defaultProject": "knora-ui",
