--- projects/knora/authentication/package.json
@@ -1,27 +1,0 @@
-{
-  "name": "@knora/authentication",
-  "version": "9.4.1",
-  "description": "Knora ui module: authentication",
-  "repository": {
-    "type": "git",
-    "url": "https://github.com/dasch-swiss/knora-ui.git"
-  },
-  "homepage": "https://dasch-swiss.github.io/knora-ui",
-  "license": "MIT",
-  "keywords": [
-    "angular",
-    "knora",
-    "digital humanities",
-    "dhlab basel"
-  ],
-  "peerDependencies": {
-    "@angular/common": "^8.0.3",
-    "@angular/core": "^8.0.3",
-    "@angular/animations": "^8.0.3",
-    "@angular/cdk": "^8.1.4",
-    "@angular/material": "^8.1.4",
-    "@knora/action": "9.4.1",
-    "@knora/core": "9.4.1",
-    "moment": "^2.22.2"
-  }
-}
