--- projects/knora/authentication/src/lib/declarations/current-user.ts
@@ -1,20 +1,0 @@
-/**
- * TODO: move into session.service
- * currently logged-in user
- */
-export interface CurrentUser {
-    // username
-    name: string;
-
-    // json web token
-    jwt: string;
-
-    // default language for ui
-    lang: string;
-
-    // is system admin?
-    sysAdmin: boolean;
-
-    // list of project shortcodes where the user is project admin
-    projectAdmin: string[];
-}
