--- projects/knora/authentication/src/lib/declarations/index.ts
@@ -1,2 +1,0 @@
-export * from './current-user';
-export * from './session';
