--- projects/knora/authentication/src/lib/declarations/session.ts
@@ -1,7 +1,0 @@
-// TODO: move into session.service
-import { CurrentUser } from './current-user';
-
-export interface Session {
-    id: number;
-    user: CurrentUser;
-}
