--- projects/knora/authentication/src/lib/interceptors/index.ts
@@ -1,3 +1,0 @@
-export * from './error.interceptor';
-export * from './jwt.interceptor';
-export * from './with-credentials.interceptor';
