--- projects/knora/authentication/src/lib/interceptors/with-credentials.interceptor.ts
@@ -1,26 +1,0 @@
-import { HttpEvent, HttpHandler, HttpInterceptor, HttpRequest } from '@angular/common/http';
-import { Injectable } from '@angular/core';
-import { Observable } from 'rxjs';
-
-import { SessionService } from '../session/session.service';
-
-@Injectable({
-    providedIn: 'root'
-})
-export class WithCredentialsInterceptor implements HttpInterceptor {
-
-    constructor(private _session: SessionService) {
-    }
-
-    intercept(request: HttpRequest<any>, next: HttpHandler): Observable<HttpEvent<any>> {
-        // add authorization header with jwt token if available
-
-        // console.log('WithCredentialsInterceptor - intercept - request: ', request);
-
-        request = request.clone({
-            withCredentials: true
-        });
-
-        return next.handle(request);
-    }
-}
