--- projects/knora/authentication/src/lib/login-form/login-form.component.html
@@ -1,59 +1,0 @@
-<div class="login-container" *ngIf="!session">
-
-    <h2 class="login-form-title mat-title">{{formLabel.title}}</h2>
-
-    <!-- This is the login form -->
-    <form [formGroup]="form" (ngSubmit)="login()" class="login-form">
-
-        <!-- Username -->
-        <mat-form-field class="full-width login-field">
-            <!-- <mat-icon matPrefix>person</mat-icon> -->
-
-            <input matInput autofocus [placeholder]="formLabel.name" autocomplete="username" formControlName="username"
-                   #username cdkFocusInitial>
-
-        </mat-form-field>
-
-        <!-- Password -->
-        <mat-form-field class="full-width login-field">
-
-            <!-- <mat-icon matPrefix>lock</mat-icon> -->
-
-            <input matInput type="password" [placeholder]="formLabel.pw" autocomplete="current-password"
-                   formControlName="password">
-
-        </mat-form-field>
-
-        <br>
-
-        <!-- Button: Login -->
-        <button mat-raised-button type="submit" [disabled]="!form.valid" class="full-width submit-button"
-                [class.mat-primary]="!errorMessage" [class.mat-warn]="errorMessage">
-            <kui-progress-indicator [color]="'white'" [status]="0" *ngIf="loading" class="login-progress">
-            </kui-progress-indicator>
-            <mat-icon *ngIf="errorMessage && !loading">close</mat-icon>
-            {{errorMessage ? formLabel.retry : formLabel.submit}}
-        </button>
-
-        <br><br>
-
-        <!-- Error message as an answer from knora api request -->
-        <p class="full-width login-error" *ngIf="errorMessage">
-            {{loginErrorServer ? formLabel.error.server : formLabel.error.failed}}
-        </p>
-    </form>
-</div>
-
-
-<!-- a user is already logged in; show who it is and a logout button -->
-<div class="logout-container" *ngIf="session">
-    <p>A user is already logged in:</p>
-    <p>Username: <strong>{{session.user.name}}</strong></p>
-    <br>
-    <p>Please log out if it's not you.</p>
-    <button mat-raised-button (click)="logout()" class="full-width mat-warn logout-button">
-        <kui-progress-indicator [color]="'white'" [status]="0" *ngIf="loading" class="login-progress">
-        </kui-progress-indicator>
-        {{formLabel.logout}}
-    </button>
-</div>
