--- projects/knora/authentication/src/lib/login-form/login-form.component.scss
@@ -1,38 +1,0 @@
-$warn: #ef5350;
-
-$width: 280px;
-
-.full-width {
-  width: 100%;
-}
-
-.login-form-title {
-  text-align: center;
-}
-
-.login-error {
-  color: $warn;
-  text-align: center;
-}
-
-.login-progress {
-  display: inline-block;
-  margin-right: 6px;
-}
-
-.login-container,
-.logout-container {
-  margin-left: auto;
-  margin-right: auto;
-  position: relative;
-  width: $width;
-}
-
-.submit-button,
-.logout-button {
-  height: 48px;
-}
-
-.center {
-  text-align: center;
-}
