--- projects/knora/authentication/src/lib/login-form/login-form.component.spec.ts
@@ -1,56 +1,0 @@
-import { async, ComponentFixture, TestBed } from '@angular/core/testing';
-import { ReactiveFormsModule } from '@angular/forms';
-import { MatButtonModule } from '@angular/material';
-import { MatFormFieldModule } from '@angular/material/form-field';
-import { MatIconModule } from '@angular/material/icon';
-import { MatInputModule } from '@angular/material/input';
-import { RouterTestingModule } from '@angular/router/testing';
-import { KuiActionModule } from '@knora/action';
-import { KuiConfigToken, KuiCoreConfig, KuiCoreModule } from '@knora/core';
-
-import { AuthenticationService } from '../authentication.service';
-import { SessionService } from '../session/session.service';
-
-import { LoginFormComponent } from './login-form.component';
-
-describe('LoginFormComponent', () => {
-    let component: LoginFormComponent;
-    let fixture: ComponentFixture<LoginFormComponent>;
-
-    beforeEach(async(() => {
-        TestBed.configureTestingModule({
-            imports: [
-                KuiActionModule,
-                KuiCoreModule,
-                MatButtonModule,
-                MatFormFieldModule,
-                MatIconModule,
-                MatInputModule,
-                ReactiveFormsModule,
-                RouterTestingModule
-            ],
-            declarations: [
-                LoginFormComponent
-            ],
-            providers: [
-                AuthenticationService,
-                SessionService,
-                {
-                    provide: KuiConfigToken,
-                    useValue: KuiCoreConfig
-                }
-            ]
-        })
-            .compileComponents();
-    }));
-
-    beforeEach(() => {
-        fixture = TestBed.createComponent(LoginFormComponent);
-        component = fixture.componentInstance;
-        fixture.detectChanges();
-    });
-
-    it('should create', () => {
-        expect(component).toBeTruthy();
-    });
-});
