--- projects/knora/authentication/src/lib/session/session.service.spec.ts
@@ -1,25 +1,0 @@
-import { HttpClient, HttpClientModule } from '@angular/common/http';
-import { inject, TestBed } from '@angular/core/testing';
-import { KuiCoreModule } from '@knora/core';
-
-import { SessionService } from './session.service';
-
-describe('SessionService', () => {
-    beforeEach(() => {
-        TestBed.configureTestingModule({
-            imports: [
-                HttpClientModule,
-                KuiCoreModule.forRoot({ name: '', api: '', app: '', media: '', ontologyIRI: '' })
-            ],
-            providers: [
-                SessionService,
-                HttpClient,
-                HttpClientModule
-            ]
-        });
-    });
-
-    it('should be created', inject([SessionService], (service: SessionService) => {
-        expect(service).toBeTruthy();
-    }));
-});
