--- projects/knora/authentication/src/public_api.ts
@@ -1,11 +1,0 @@
-/*
- * Public API Surface of authentication
- */
-
-// export * from './lib/interceptors/';
-// export * from './lib/declarations/';
-export * from './lib/guard/auth.guard';
-export * from './lib/login-form/login-form.component';
-export * from './lib/session/session.service';
-// export * from './lib/authentication.service';
-export * from './lib/authentication.module';
