--- projects/knora/authentication/tslint.json
@@ -1,17 +1,0 @@
-{
-    "extends": "../../../tslint.json",
-    "rules": {
-        "directive-selector": [
-            true,
-            "attribute",
-            "kui",
-            "camelCase"
-        ],
-        "component-selector": [
-            true,
-            "element",
-            "kui",
-            "kebab-case"
-        ]
-    }
-}
