+++ src/app/app.routing.ts
@@ -1,11 +1,9 @@
 import { NgModule } from '@angular/core';
 import { RouterModule, Routes } from '@angular/router';
-
 import { ResourceViewComponent } from '@knora/viewer';
-import { AppDemo } from './app.config';
-// dev docs
-import { DocIntroComponent } from './dev-docs/doc-intro/doc-intro.component';
 
+import { AppDemo } from './app.config';
+import { DocIntroComponent } from './dev-docs/doc-intro/doc-intro.component';
 import { ActionDemoComponent } from './knora-ui-examples/action-demo/action-demo.component';
 import { AdminImageComponent } from './knora-ui-examples/action-demo/admin-image/admin-image.component';
 import { ExistingNameComponent } from './knora-ui-examples/action-demo/existing-name/existing-name.component';
@@ -13,10 +11,9 @@ import { KeyComponent } from './knora-ui-examples/action-demo/key/key.component'
 import { MessageComponent } from './knora-ui-examples/action-demo/message/message.component';
 import { ProgressIndicatorComponent } from './knora-ui-examples/action-demo/progress-indicator/progress-indicator.component';
 import { SortButtonComponent } from './knora-ui-examples/action-demo/sort-button/sort-button.component';
-import { AuthComponent } from './knora-ui-examples/authentication-demo/auth/auth.component';
-import { AuthenticationDemoComponent } from './knora-ui-examples/authentication-demo/authentication-demo.component';
-
-import { LoginComponent } from './knora-ui-examples/authentication-demo/login/login.component';
+import { StringifyStringLiteralComponent } from './knora-ui-examples/action-demo/stringify-string-literal/stringify-string-literal.component';
+import { TruncateComponent } from './knora-ui-examples/action-demo/truncate/truncate.component';
+import { LoginComponent } from './knora-ui-examples/action-demo/login/login.component';
 import { ConvertJsonLdComponent } from './knora-ui-examples/core-demo/convert-json-ld/convert-json-ld.component';
 import { CoreDemoComponent } from './knora-ui-examples/core-demo/core-demo.component';
 import { GravsearchComponent } from './knora-ui-examples/core-demo/gravsearch/gravsearch.component';
@@ -29,27 +26,19 @@ import { ProjectsComponent } from './knora-ui-examples/core-demo/projects/projec
 import { ResourceComponent } from './knora-ui-examples/core-demo/resource/resource.component';
 import { SearchComponent } from './knora-ui-examples/core-demo/search/search.component';
 import { UsersComponent } from './knora-ui-examples/core-demo/users/users.component';
-
-import { SearchDemoComponent } from './knora-ui-examples/search-demo/search-demo.component';
-import { SearchPanelComponent } from './knora-ui-examples/search-demo/search-panel/search-panel.component';
 import { ExtendedSearchComponent } from './knora-ui-examples/search-demo/extended-search/extended-search.component';
 import { FulltextSearchComponent } from './knora-ui-examples/search-demo/fulltext-search/fulltext-search.component';
-
+import { SearchDemoComponent } from './knora-ui-examples/search-demo/search-demo.component';
+import { SearchPanelComponent } from './knora-ui-examples/search-demo/search-panel/search-panel.component';
 import { PropertiesComponent } from './knora-ui-examples/viewer-demo/properties/properties.component';
 import { ResourcesComponent } from './knora-ui-examples/viewer-demo/resources/resources.component';
-
+import { SearchResultsComponent } from './knora-ui-examples/viewer-demo/search-results/search-results.component';
 import { ViewerDemoComponent } from './knora-ui-examples/viewer-demo/viewer-demo.component';
 import { ViewsComponent } from './knora-ui-examples/viewer-demo/views/views.component';
-import { SearchResultsComponent } from './knora-ui-examples/viewer-demo/search-results/search-results.component';
-
 import { DemoIntroComponent } from './landing-page/demo-intro/demo-intro.component';
-// / start with main-intro
 import { MainIntroComponent } from './landing-page/main-intro/main-intro.component';
 import { TreeComponent } from './material/tree/tree.component';
-// /modules demo
 import { ModuleIndexComponent } from './partials/module-index/module-index.component';
-import { TruncateComponent } from './knora-ui-examples/action-demo/truncate/truncate.component';
-import { StringifyStringLiteralComponent } from './knora-ui-examples/action-demo/stringify-string-literal/stringify-string-literal.component';
 
 const appRoutes: Routes = [
     {
@@ -89,6 +78,10 @@ const appRoutes: Routes = [
                         component: SortButtonComponent
                     },
                     {
+                        path: 'login-form',
+                        component: LoginComponent
+                    },
+                    {
                         path: 'message',
                         component: MessageComponent
                     },
@@ -116,31 +109,6 @@ const appRoutes: Routes = [
                         path: 'stringify-string-literal',
                         component: StringifyStringLiteralComponent
                     }
-                ]
-            },
-            {
-                path: 'authentication',
-                component: AuthenticationDemoComponent,
-                // canActivate: [AuthGuard],
-                children: [
-                    {
-                        path: '',
-                        redirectTo: 'readme',
-                        pathMatch: 'full'
-                    },
-                    {
-                        path: 'readme',
-                        component: DemoIntroComponent,
-                        data: { module: 'authentication' }
-                    },
-                    {
-                        path: 'login-form',
-                        component: LoginComponent
-                    },
-                    {
-                        path: 'authentication',
-                        component: AuthComponent
-                    },
                 ]
             },
             {
