+++ src/app/knora-ui-examples/action-demo/login/login.component.scss
