+++ src/app/knora-ui-examples/action-demo/login/login.component.ts
@@ -1,4 +1,5 @@
 import { Component, OnInit } from '@angular/core';
+
 import { AppDemo } from '../../../app.config';
 import { Example } from '../../../app.interfaces';
 
@@ -9,12 +10,12 @@ import { Example } from '../../../app.interfaces';
 })
 export class LoginComponent implements OnInit {
 
-    module = AppDemo.authenticationModule;
+    module = AppDemo.actionModule;
 
     example: Example = {
         title: 'kui-login-form',
         subtitle: 'Creates a login form, if the user is not yet logged-in. Otherwise it shows a logout button',
-        name: 'loginForm',
+        name: 'loginform',
         code: {
             html: `
 <kui-login-form></kui-login-form>`,
