--- src/app/knora-ui-examples/authentication-demo/auth/auth.component.html
@@ -1,2 +1,0 @@
-<!-- Documentation from JSDocs -->
-<app-documentation-viewer [module]="module"></app-documentation-viewer>
