--- src/app/knora-ui-examples/authentication-demo/authentication-demo.component.html
@@ -1,1 +1,0 @@
-<router-outlet></router-outlet>
