--- src/app/knora-ui-examples/authentication-demo/authentication-demo.component.ts
@@ -1,15 +1,0 @@
-import { Component, OnInit } from '@angular/core';
-
-@Component({
-  selector: 'app-authentication-demo',
-  templateUrl: './authentication-demo.component.html',
-  styleUrls: ['./authentication-demo.component.scss']
-})
-export class AuthenticationDemoComponent implements OnInit {
-
-  constructor() { }
-
-  ngOnInit() {
-  }
-
-}
